requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
ImageHash==4.3.1
requests==2.32.3
